Make analyte test schema validated, switch back GQN property name
