+++ src/encoded/schemas/analyte.json
@@ -30,17 +30,17 @@
         "dna_integrity_number_instrument": [
             "dna_integrity_number"
         ],
-        "dna_genomic_quality_number": [
-            "dna_genomic_quality_number_instrument",
-            "dna_genomic_quality_size_threshold"
+        "genomic_quality_number": [
+            "genomic_quality_number_instrument",
+            "genomic_quality_size_threshold"
         ],
-        "dna_genomic_quality_number_instrument": [
-            "dna_genomic_quality_number",
-            "dna_genomic_quality_size_threshold"
+        "genomic_quality_number_instrument": [
+            "genomic_quality_number",
+            "genomic_quality_size_threshold"
         ],
-        "dna_genomic_quality_size_threshold": [
-            "dna_genomic_quality_number",
-            "dna_genomic_quality_number_instrument"
+        "genomic_quality_size_threshold": [
+            "genomic_quality_number",
+            "genomic_quality_number_instrument"
         ],
         "dna_quality_number": [
             "dna_quality_number_instrument",
@@ -193,14 +193,14 @@
             "type": "number",
             "minimum": 0
         },
-        "dna_genomic_quality_number": {
+        "genomic_quality_number": {
             "title": "Genomic Quality Number",
             "description": "Genomic quality number, percent of samples above threshold value",
             "type": "number",
             "minimum": 0,
             "maximum": 10
         },
-        "dna_genomic_quality_number_instrument": {
+        "genomic_quality_number_instrument": {
             "title": "Genomic Quality Number Instrument",
             "description": "Instrument used for GQN assessment",
             "type": "array",
@@ -212,7 +212,7 @@
                 ]
             }
         },
-        "dna_genomic_quality_size_threshold": {
+        "genomic_quality_size_threshold": {
             "title": "Genomic Quality Size Threshold",
             "description": "Threshold set in ProSize for genomic quality number, for assessing percent of samples above threshold size (bp), if applicable",
             "type": "number",
@@ -368,6 +368,45 @@
             ]
         }
     },
+    "anyOf": [
+        {
+            "properties": {
+                "molecule": {
+                    "not": {
+                        "contains": {
+                            "const": "DNA"
+                        }
+                    }
+                }
+            },
+            "allOf": [
+                {"not": { "required": ["dna_genomic_quality_number"]}},
+                {"not": { "required": ["dna_genomic_quality_number_instrument"]}},
+                {"not": { "required": ["dna_genomic_quality_size_threshold"]}},
+                {"not": { "required": ["dna_integrity_number"]}},
+                {"not": { "required": ["dna_integrity_number_instrument"]}},
+                {"not": { "required": ["dna_quality_number"]}},
+                {"not": { "required": ["dna_quality_number_instrument"]}},
+                {"not": { "required": ["dna_quality_size_threshold"]}}
+            ]
+        },
+        {
+            "properties": {
+                "molecule": {
+                    "not": {
+                        "contains": {
+                            "const": "RNA"
+                        }
+                    }
+                }
+            },
+            "allOf": [
+                {"not": { "required": ["rna_integrity_number"]}},
+                {"not": { "required": ["rna_integrity_number_instrument"]}},
+                {"not": { "required": ["ribosomal_rna_ratio"]}}
+            ]
+        }
+    ],
     "facets": {
         "submission_centers.display_title": {
             "title": "Submitted By"
