+++ src/encoded/tests/test_schema_analyte.py
@@ -14,20 +14,23 @@ from ..item_utils import (
 
 @pytest.mark.workbook
 @pytest.mark.parametrize(
-    "patch_body,expected_status",
+    "molecule,patch_body,expected_status",
     [
-        ({"rna_integrity_number": 6}, 422),
-        ({"rna_integrity_number_instrument": "Agilent Bioanalyzer"}, 422),
-        (
+        ("RNA", {"rna_integrity_number": 6}, 422),
+        ("RNA", {"rna_integrity_number_instrument": "Agilent Bioanalyzer"}, 422),
+        ("RNA",
             {
                 "rna_integrity_number_instrument": "Agilent Bioanalyzer",
                 "rna_integrity_number": 6,
             },
             200,
         ),
+        ("DNA", {"rna_integrity_number_instrument": "Agilent Bioanalyzer"}, 422),
+
     ],
 )
 def test_rna_integrity_requirements(
+    molecule: str,
     patch_body: Dict[str, Any],
     expected_status: int,
     es_testapp: TestApp,
@@ -38,7 +41,7 @@ def test_rna_integrity_requirements(
         get_item_from_search(
             es_testapp,
             "Analyte",
-            add_on="&molecule=RNA"
+            add_on=f"&molecule={molecule}"
         )
     )
     patch_item(
@@ -51,20 +54,23 @@ def test_rna_integrity_requirements(
 
 @pytest.mark.workbook
 @pytest.mark.parametrize(
-    "patch_body,expected_status",
+    "molecule,patch_body,expected_status",
     [
-        ({"dna_integrity_number": 6}, 422),
-        ({"dna_integrity_number_instrument": "Agilent 4200 TapeStation"}, 422),
-        (
+        ("DNA", {"dna_integrity_number": 6}, 422),
+        ("DNA", {"dna_integrity_number_instrument": "Agilent 4200 TapeStation"}, 422),
+        ("DNA",
             {
                 "dna_integrity_number": 6,
                 "dna_integrity_number_instrument": "Agilent 4200 TapeStation",
             },
             200,
         ),
+        ("RNA", {"dna_integrity_number_instrument": "Agilent 4200 TapeStation"}, 422),
+
     ],
 )
 def test_dna_integrity_requirements(
+    molecule: str,
     patch_body: Dict[str, Any],
     expected_status: int,
     es_testapp: TestApp,
@@ -75,7 +81,7 @@ def test_dna_integrity_requirements(
         get_item_from_search(
             es_testapp,
             "Analyte",
-            add_on="&molecule=DNA"
+            add_on=f"&molecule={molecule}"
         )
     )
     patch_item(
@@ -88,26 +94,26 @@ def test_dna_integrity_requirements(
 
 @pytest.mark.workbook
 @pytest.mark.parametrize(
-    "patch_body,expected_status",
+    "molecule,patch_body,expected_status",
     [
-        ({"dna_quality_number": 6}, 422),
-        ({"dna_quality_number_instrument": "Agilent 5400 Fragment Analyzer"}, 422),
-        ({"dna_quality_size_threshold": 1000}, 422),
-        (
+        ("DNA", {"dna_quality_number": 6}, 422),
+        ("DNA", {"dna_quality_number_instrument": "Agilent 5400 Fragment Analyzer"}, 422),
+        ("DNA", {"dna_quality_size_threshold": 1000}, 422),
+        ("DNA", 
             {
                 "dna_quality_number": 6,
                 "dna_quality_number_instrument": "Agilent 4200 TapeStation",
             },
             422,
         ),
-        (
+        ("DNA", 
             {
                 "dna_quality_number": 6,
                 "dna_quality_size_threshold": 1000,
             },
             422,
         ),
-        (
+        ("DNA", 
             {
                 "dna_quality_number": 6,
                 "dna_quality_size_threshold": 1000,
@@ -115,9 +121,18 @@ def test_dna_integrity_requirements(
             },
             200,
         ),
+        ("RNA", 
+            {
+                "dna_quality_number": 6,
+                "dna_quality_size_threshold": 1000,
+                "dna_quality_number_instrument": "Agilent 4200 TapeStation",
+            },
+            422,
+        ),
     ],
 )
 def test_dna_quality_requirements(
+    molecule: str,
     patch_body: Dict[str, Any],
     expected_status: int,
     es_testapp: TestApp,
@@ -128,7 +143,7 @@ def test_dna_quality_requirements(
         get_item_from_search(
             es_testapp,
             "Analyte",
-            add_on="&molecule=DNA"
+            add_on=f"&molecule={molecule}"
         )
     )
     patch_item(
@@ -141,39 +156,48 @@ def test_dna_quality_requirements(
 
 @pytest.mark.workbook
 @pytest.mark.parametrize(
-    "patch_body,expected_status",
+    "molecule,patch_body,expected_status",
     [
-        ({"dna_genomic_quality_number": 6}, 422),
-        (
-            {"dna_genomic_quality_number_instrument": ["Agilent 5400 Fragment Analyzer"]},
+        ("DNA", {"genomic_quality_number": 6}, 422),
+        ("DNA",
+            {"genomic_quality_number_instrument": ["Agilent 5400 Fragment Analyzer"]},
             422,
         ),
-        ({"dna_genomic_quality_size_threshold": 1000}, 422),
-        (
+        ("DNA", {"genomic_quality_size_threshold": 1000}, 422),
+        ("DNA",
             {
-                "dna_genomic_quality_number": 6,
-                "dna_genomic_quality_number_instrument": ["Agilent 4200 TapeStation"],
+                "genomic_quality_number": 6,
+                "genomic_quality_number_instrument": ["Agilent 4200 TapeStation"],
             },
             422,
         ),
-        (
+        ("DNA",
             {
-                "dna_genomic_quality_number": 6,
-                "dna_genomic_quality_size_threshold": 1000,
+                "genomic_quality_number": 6,
+                "genomic_quality_size_threshold": 1000,
             },
             422,
         ),
-        (
+        ("DNA",
             {
-                "dna_genomic_quality_number": 6,
-                "dna_genomic_quality_size_threshold": 1000,
-                "dna_genomic_quality_number_instrument": ["Agilent 4200 TapeStation"],
+                "genomic_quality_number": 6,
+                "genomic_quality_size_threshold": 1000,
+                "genomic_quality_number_instrument": ["Agilent 4200 TapeStation"],
             },
             200,
         ),
+        ("RNA",
+            {
+                "genomic_quality_number": 6,
+                "genomic_quality_size_threshold": 1000,
+                "genomic_quality_number_instrument": ["Agilent 4200 TapeStation"],
+            },
+            422,
+        ),
     ],
 )
-def test_dna_genome_quality_requirements(
+def test_genome_quality_requirements(
+    molecule: str,
     patch_body: Dict[str, Any],
     expected_status: int,
     es_testapp: TestApp,
@@ -184,7 +208,7 @@ def test_dna_genome_quality_requirements(
         get_item_from_search(
             es_testapp,
             "Analyte",
-            add_on="&molecule=DNA"
+            add_on=f"&molecule={molecule}"
         )
     )
     patch_item(
@@ -193,3 +217,4 @@ def test_dna_genome_quality_requirements(
         dna_uuid,
         status=expected_status,
     )
+
